clean-up of IMD TTD/MD data routines

+++ src/particles/ttm/ttm_vars.f90
@@ -19,7 +19,7 @@ REAL,ALLOCATABLE,TARGET    :: TTM_DG(:,:,:,:,:)                                 
 ! TTM-FD solution
 REAL,ALLOCATABLE,TARGET    :: TTM_FD(:,:,:,:)                                   !> TTM FD bary center solution
 REAL,ALLOCATABLE,TARGET    :: ElemBaryFD(:,:)                                   !> TTM FD bary center position
-REAL,ALLOCATABLE,TARGET    :: ElemIndexFD(:,:)                                  !> TTM FD index position of FD cell in DG grid
+INTEGER,ALLOCATABLE,TARGET :: ElemIndexFD(:,:)                                  !> TTM FD index position of FD cell in DG grid
 REAL                       :: TTMElemBaryTolerance                              !> TTM FD bary center tolerance to DG bary center
 LOGICAL,ALLOCATABLE,TARGET :: ElemIsDone(:)                                     !> TTM FD bary center has been found
 INTEGER                    :: TTMGridFDdim(3)                                   !> number of FD grid cells in each direction
